Renaming orm model module to reflect they are orm models
