+++ backend/app/repository/models/__init__.py
@@ -1,5 +1,5 @@
 from .base import BaseSQLModel
-from .yugioh_card import YugiohCardORM
-from .yugioh_set import YugiohSetORM, YugiohCardSetAssociation
+from .yugioh_card_orm import YugiohCardORM
+from .yugioh_set_orm import YugiohSetORM, YugiohCardSetAssociation
 
 __all__ = ["BaseSQLModel", "YugiohCardORM", "YugiohSetORM", "YugiohCardSetAssociation"]
